requirements.txt
aiohttp==3.8.3 # HTTP client
pathy==0.10.1 # Path manipulation
pydantic==2.9 # Data validation
pydantic[email,timezone]==2.9 # Data validation
PyYAML==6.0 # YAML parser
python-dotenv==1.0.0 # Environment variables
python-socketio==5.8.0 # Socket.IO
regex==2022.3.15 # Regular expressions
requests==2.31.0 # HTTP client
requests-oauthlib==1.3.1 # OAuth
psycopg==3.1.9 # PostgreSQL
psycopg-binary==3.1.9 # PostgreSQL
sanic==24.6.0 # Sanic
Sanic_Cors==2.2.0 # Sanic CORS
sanic-routing==23.12.0 # Sanic Router 
sanic-ext==23.12.0
sanic_openapi==21.12.0
pymongo==4.8.0 # MongoDB
# pylint==3.3.0 # Linter
aio_pika==7.0.0 # RabbitMQ
motor==3.5.1 # MongoDB
firebase-admin==6.5.0 # Firebase
apscheduler==3.10.4 # Scheduler
#infobip-api-python-sdk==5.0.1 # SMS API
aiosmtplib==3.0.2 # SMTP client 
aioimaplib==1.1.0 # IMAP client
beautifulsoup4==4.12.3 # HTML parser
ruamel.yaml==0.18.6 # YAML parser
flake8==7.1.1 # Linter
black==24.8.0 # Code formatter
autoflake==2.3.1 # Code formatter
pytest==8.2.0 # Testing
coverage==7.6.1 # Code coverage
sanic-testing==23.12.0 # Sanic testing
redis==5.1.0 # Redis
structlog==24.4.0 # Logging